+++ examples/other/tts_finetune/tts3/conf/finetune.yaml
@@ -9,4 +9,6 @@ num_snapshots: -1
 
 # frozen_layers should be a list
 # if you don't need to freeze, set frozen_layers to []
+# fastspeech2 layers can be found on conf/fastspeech2_layers.txt
+# example: frozen_layers: ["encoder", "duration_predictor"]
 frozen_layers: ["encoder"]
